@@ -1,4 +1,5 @@
-import mkdocs
+from typing import override
+from mkdocs.plugins import BasePlugin
 from mkdocs.config.defaults import MkDocsConfig
 from mkdocs.structure.pages import Page
 from mkdocs.structure.files import Files
@@ -6,6 +7,8 @@ from mkdocs.structure.files import Files
 from .on_page_markdown import on_page_markdown
 from .config import EntangledConfig
 
-class EntangledPlugin(mkdocs.plugins.BasePlugin[EntangledConfig]):
+
+class EntangledPlugin(BasePlugin[EntangledConfig]):
+    @override
     def on_page_markdown(self, markdown: str, *, page: Page, config: MkDocsConfig, files: Files):
         return on_page_markdown(markdown, page=page, config=config, files=files)
